% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfinference.R
\name{borehole}
\alias{borehole}
\title{borehole}
\usage{
borehole(xx)
}
\arguments{
\item{xx}{a design matrix of 8 dimension}
}
\value{
a vector containing evaluated values of borehole function.
}
\description{
a test function
}
\examples{
# borehole(xx)

}
