% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfinference.R
\name{cond_root}
\alias{cond_root}
\title{cond_root}
\usage{
cond_root(tree, id)
}
\arguments{
\item{tree}{a single tree object obtained from using \code{"randomForest"} and \code{"getTree"}}

\item{id}{index of the tree}
}
\value{
a statement defining the leaf (associated with "id"), by collecting the entire "ancestry"
}
\description{
finding the "parents"
using this function recursively, we can find the family ancestry.
}
\examples{
# cond_root(tree, id)

}
