% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfinference.R
\name{getConds}
\alias{getConds}
\title{getConds}
\usage{
getConds(tree)
}
\arguments{
\item{tree}{a single tree object obtained from using \code{"randomForest"} and \code{"getTree"}}
}
\value{
a statement defining the leaf (associated with "id"), by collecting the entire "ancestry"
}
\description{
give the partition statements for each leaf; a wrapper using cond_root and other functions
included for backward compatibility
}
\details{
using this function recursively, we can find the family ancestry.
}
\examples{
# getConds(tree)

}
