% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfinference.R
\name{id_root}
\alias{id_root}
\title{id_root}
\usage{
id_root(tree, id)
}
\arguments{
\item{tree}{a single tree object obtained from using \code{"randomForest"} and \code{"getTree"}}

\item{id}{index of the tree}
}
\value{
a vector containing the entire "ancestry"
}
\description{
finding the "parents"; recovering the sequence of node ids from the given leaf id.
using prevCond function recursively, we can find the family ancestry
}
\examples{
# id_root(tree, id)

}
