% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfinference.R
\name{prevCond}
\alias{prevCond}
\title{finding the previous conditions in the tree}
\usage{
prevCond(tree, i)
}
\arguments{
\item{tree}{a single tree object obtained from using \code{"randomForest"} and \code{"getTree"}}

\item{i}{index of the tree}
}
\value{
a list containing \code{"cond"} and \code{"id"})
}
\description{
finding the "parents"
using this function recursively, we can find the family ancestry.
}
\examples{
# prevCond(tree, i)

}
